requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiogram>=3.15.0
aiohttp>=3.11.0
yt-dlp>=2025.1.0
python-dotenv>=1.0.0

# ── System dependency (not pip) ──
# ffmpeg required for Spotify playback (YouTube → MP3 conversion)
# Ubuntu/Debian: sudo apt install ffmpeg
# macOS:         brew install ffmpeg
# Windows:       https://ffmpeg.org/download.html